rpmostreepayload: Do /sysroot mount non-recursively
In https://github.com/rhinstaller/anaconda/pull/903 we
changed all of rpmostreepayload's internal mounts to be
recursive to fix `/boot/efi`.  And indeed all should be recursive,
except one - the `/sysroot` mount.  By asking for that to be
recursive we've created a loop.

I noticed a while ago that the output of `findmnt` during an install was insane
(the deployment directory was visible twice, etc.), but never debugged it. It
turns out this was the cause.

Fixing this makes the mounts look a lot more sane, and will help
for future work I have for `/var` mounts.