rpmostreepayload: Fix logic for copying of EFI data
We were copying the `efi` dir even on BIOS systems - the code was intending to
avoid this but wasn't doing so. We actually need to fix things so that we detect
whether an EFI-based bootloader is in use, rather than whether or not the
directory already existed (it can't have anyways since nothing would copy data
there yet).